Remove Clang as default compiler
Specifying the default like that was probably bad practice. That's
especially true of the explicit and now dated version that was
specified.

Code should now build with the environment default compiler, which is
typically GCC. It can be modified by the CXX environment variable. This
can be done directly in the make command.

Example:
```
make CXX=g++
```

Example:
```
make CXX=clang++
```